optimize db and fix parser #262

+++ OpenNos.DAL.EF.MySQL/DB/ShopItem.cs
@@ -16,11 +16,11 @@ namespace OpenNos.DAL.EF.MySQL.DB
     {
         public int ShopItemId { get; set; }
         public byte Type { get; set; }
-        public short Slot { get; set; }
+        public byte Slot { get; set; }
         public short ItemVNum { get; set; }
         public byte Upgrade { get; set; }
         public byte Rare { get; set; }
-        public short Color { get; set; }
+        public byte Color { get; set; }
         public int ShopId { get; set; }
     
         public virtual Item item { get; set; }
