+++ OpenNos.Data/MapMonsterDTO.cs
@@ -23,7 +23,7 @@ namespace OpenNos.Data
         public short MapId { get; set; }
         public short MapX { get; set; }
         public short MapY { get; set; }
-        public short Position { get; set; }
+        public byte Position { get; set; }
         public bool Move { get; set; }
 
         #endregion
